+++ src/main/scala/holes/HoleName.scala
@@ -1,0 +1,3 @@
+package holes
+
+case class HoleName(name: String)
